Move validation related types to validators.d.ts
