+++ src/types/index.d.ts
@@ -1,2 +1,3 @@
 export type * from './models.d.ts';
 export type * from './pies.d.ts';
+export type * from './validators.d.ts';
